requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
psycopg2==2.9.6
Werkzeug==2.3.7
python-dotenv==1.0.0
reportlab==4.0.4